feat(habitat): move check baseline ownership into Effect services
Move check and baseline implementation ownership out of src/lib into structural-check and baseline-authority domains, add BaselineAuthority and StructuralCheck Effect services, and wire the shared Habitat runtime so direct Effect programs and Effect-oRPC service modules receive the same domain layers. Service procedure logic now lives directly in each service module router instead of imported run.ts wrappers, matching the Effect-oRPC router pattern.

Validation: bun run biome:ci; bun run --cwd tools/habitat-harness check; bun run --cwd tools/habitat-harness test -- test/service/service-architecture.test.ts test/service/check-service.test.ts test/service/classify-service.test.ts test/service/fix-service.test.ts test/service/graph-service.test.ts test/service/hook-service.test.ts test/lib/verify-service.test.ts test/lib/hooks.test.ts test/lib/pattern-apply.test.ts test/lib/rule-selection.test.ts test/service/transactions-service.test.ts; bun run openspec -- validate deep-habitat-effect-check-baseline-cutover --strict; bun run openspec:validate; git diff --check.

Current-tree pattern-check remains the known next repair target when Grit exits 130 on source-rule execution.
