+++ tools/habitat-harness/src/domains/baseline-authority/application.ts
@@ -1,4 +1,4 @@
-import type { HabitatDiagnostic } from "../diagnostics.js";
+import type { HabitatDiagnostic } from "../../lib/diagnostics.js";
 import {
   type BaselineApplicationResult,
   type BaselineAuthorityState,
