+++ tools/habitat-harness/src/domains/structural-check/disposition-diagnostics.ts
