+++ tools/habitat-harness/src/domains/structural-check/schema.ts
