+++ tools/habitat-harness/src/domains/structural-check/state.ts
