+++ tools/habitat-harness/src/service/impl.ts
@@ -1,11 +1,15 @@
 import { Layer, ManagedRuntime } from "effect";
 import { type EffectImplementer, implementEffect } from "effect-orpc";
+import { BaselineAuthorityLive } from "../domains/baseline-authority/index.js";
+import { StructuralCheckLive } from "../domains/structural-check/index.js";
 import { HabitatRuntimeLive } from "../runtime/layers.js";
 import { type HabitatServiceContext, HabitatServiceRuntime } from "./base.js";
 import { habitatServiceContract } from "./contract.js";
 
 export const habitatServiceLayer = Layer.mergeAll(
   HabitatRuntimeLive,
+  BaselineAuthorityLive,
+  StructuralCheckLive,
   Layer.succeed(HabitatServiceRuntime, { service: "habitat" })
 );
