+++ tools/habitat-harness/src/service/modules/classify/router.ts
@@ -1,5 +1,8 @@
+import { Effect } from "effect";
+import { classifyTargetResult } from "../../../lib/classify-core/index.js";
+import type { ClassifyServiceOptions } from "./context.js";
+import type { ClassifyServiceRunInput } from "./contract.js";
 import { module as classifyModule } from "./module.js";
-import { runClassifyService } from "./run.js";
 
 export const classifyRouter = {
   run: classifyModule.run.effect(({ context, input }) =>
@@ -8,3 +11,10 @@ export const classifyRouter = {
 };
 
 export const router = classifyRouter;
+
+export function runClassifyService(
+  input: ClassifyServiceRunInput,
+  serviceOptions: ClassifyServiceOptions = {}
+) {
+  return Effect.promise(() => classifyTargetResult(input.target, serviceOptions.options));
+}
