+++ tools/habitat-harness/test/service/hook-service.test.ts
@@ -2,7 +2,7 @@ import { Effect } from "effect";
 import { describe, expect, test } from "vitest";
 import type { HookRuntime } from "../../src/lib/hook-runtime/runtime.js";
 import { createHabitatServiceClient } from "../../src/service/client.js";
-import { runHookService } from "../../src/service/modules/hook/run.js";
+import { runHookService } from "../../src/service/modules/hook/router.js";
 
 describe("Habitat hook service", () => {
   test("runs owned hook orchestration from service input", async () => {
